fix(ci): 🩹 give the runner the headers the shim's build script needs
`just lint` runs with --all-features, and the justfile promised that needed
neither libva nor a GPU because clippy analyses without linking. The libavcodec
shim broke that promise: a build script compiles real C and wants real headers,
whether or not anything is later linked. CI went red on the commit that added it.

Installing the -dev packages is the honest fix. The alternative — letting the
build script skip the shim when pkg-config comes up empty — would report a lint
pass over code nobody compiled, which is the failure mode this milestone keeps
being bitten by.

+++ core/crates/encoder/src/vulkan/sys.rs
@@ -33,8 +33,8 @@ pub fn device_number(node: &Path) -> Result<DeviceNumber, EncoderError> {
     })?;
     let rdev = metadata.rdev();
     Ok(DeviceNumber {
-        major: u64::from(nix::sys::stat::major(rdev)),
-        minor: u64::from(nix::sys::stat::minor(rdev)),
+        major: nix::sys::stat::major(rdev),
+        minor: nix::sys::stat::minor(rdev),
     })
 }
 
@@ -63,7 +63,10 @@ mod tests {
         let null = device_number(Path::new("/dev/null")).unwrap();
         let zero = device_number(Path::new("/dev/zero")).unwrap();
         assert_eq!(null.major, zero.major, "both are mem devices");
-        assert_ne!(null, zero, "they differ in the minor and must not compare equal");
+        assert_ne!(
+            null, zero,
+            "they differ in the minor and must not compare equal"
+        );
     }
 
     #[test]
